Add worktime command
The command reads a file containing date and login-logout times,
and prints the daily work time, and also a summary.
It's prepared for a strict format:
YYYY-MM-DD HH:MM-HH:MM HH:MM-HH:MM HH:MM-

where YYYY-MM-DD like 2012-12-31, the date of the entry
The HH:MM-HH:MM is an interval, from-to. It can be multiple.
If the last "to" time is missing, it is filled with current time.
If it's missing from an earlier date, it's an error logically,
but it's ignored, and also the current time is used.

The file name can be:
* <parameter> (if specified, remaining hardcoded names are ignored).
* ~/WT.txt
* ~/WORKTIME.txt
in this order and the first existing will be used.