fetchcovidhu: use urllib3 with retries, sleep 0.1 between requests
The host can close the connection for some reason, and it's better to
reuse the HTTP connections. Furthermore retrying the connection is also
useful, so the code is updated to use urllib3.

The pool should be cleared, otherwise similar issue occurs:

.../fetchcovidhu/fetch.py:299: ResourceWarning: unclosed
<ssl.SSLSocket fd=3, family=AddressFamily.AF_INET,
 type=SocketKind.SOCK_STREAM, proto=6, ...>
  return _Fetcher(self.directory, self.timestamp).fetch()
ResourceWarning: Enable tracemalloc to get the object allocation traceback